Final patch: fix gevent and eventlet imports
The imports for gevent and eventlet were a bit off, shall we say. This
patch fixes them to import the getcurrent methods properly.

We also add the keyword, "greenlets", as an synonym for eventlet.

NOTE to Ned, current maintainer of Coverage: We might be able to just
import the greenlets module directly, instead of handling gevent and
eventlet separately as the patch currently maintains. I am not sure if
both gevent and eventlet always use greenlets, so for now they are
separate.

Future: work with other frameworks that might do the same or similar
things, like Stackless, and be sure other event loop mechanims work
(perhaps Tulip, http://www.python.org/dev/peps/pep-3156/).

Signed-off-by: Peter Portante <peter.portante@redhat.com>